refactor: merge records page into home page
- Remove separate records page, use index as the main records view
- Home page now shows daily records with date selector
- Reduce tabBar from 4 to 3 items (记录, 历史, 设置)
- Keep custom navigation bar with logo on home page (fixed position)
- Add collapsible record cards with state persisted to local storage
