+++ src/pages/index/index.config.ts
@@ -1,3 +1,4 @@
 export default definePageConfig({
   navigationStyle: "custom",
+  enablePullDownRefresh: true,
 });
