+++ src/pages/index/index.css
@@ -1,15 +1,21 @@
-.home-page {
+.index-page {
   min-height: 100vh;
   background-color: #f5f5f5;
+  padding-bottom: 40px;
 }
 
 /* 顶部栏 */
 .top-header {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
   display: flex;
   align-items: center;
   padding: 0 24px;
   background-color: #fff;
   box-sizing: content-box;
+  z-index: 100;
 }
 
 .app-logo {
@@ -25,43 +31,109 @@
   color: #333;
 }
 
-/* 快速操作 */
-.quick-actions {
-  padding: 40px 32px;
-}
-
-.section-title {
-  font-size: 32px;
-  font-weight: 600;
-  color: #333;
-  margin-bottom: 32px;
-  display: block;
-}
-
-.action-grid {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
+/* 日期选择器 */
+.date-selector {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 24px 32px;
+  background-color: #fff;
   gap: 24px;
 }
 
-.action-item {
+.date-arrow {
+  font-size: 36px;
+  color: #07c160;
+  padding: 16px;
+}
+
+.date-arrow.disabled {
+  color: #ccc;
+}
+
+.date-text {
+  font-size: 36px;
+  font-weight: 600;
+  color: #333;
+}
+
+/* 加载状态 */
+.loading {
   display: flex;
-  flex-direction: column;
-  align-items: center;
   justify-content: center;
-  padding: 48px 24px;
+  align-items: center;
+  padding: 80px;
+  color: #999;
+  font-size: 28px;
+}
+
+/* 记录容器 */
+.records-container {
+  padding: 20px 24px;
+}
+
+/* 记录卡片 */
+.record-card {
   background-color: #fff;
   border-radius: 16px;
-  box-shadow: 0 2px 8px rgb(0 0 0 / 5%);
+  margin-bottom: 20px;
+  overflow: hidden;
 }
 
-.action-icon {
-  font-size: 56px;
-  margin-bottom: 16px;
+.card-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 24px 28px;
+  border-bottom: 1px solid #f0f0f0;
 }
 
-.action-label {
+.card-title-row {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+}
+
+.card-arrow {
+  font-size: 24px;
+  color: #999;
+  transition: transform 0.2s;
+}
+
+.card-arrow.collapsed {
+  transform: rotate(-90deg);
+}
+
+.card-icon {
+  font-size: 36px;
+}
+
+.card-title {
   font-size: 30px;
+  font-weight: 600;
   color: #333;
-  font-weight: 500;
+}
+
+.card-count {
+  font-size: 24px;
+  color: #999;
+}
+
+.card-add-btn {
+  font-size: 40px;
+  color: #07c160;
+  padding: 8px 16px;
+}
+
+.card-content {
+  padding: 20px 28px;
+  min-height: 60px;
+}
+
+.empty-hint {
+  font-size: 26px;
+  color: #ccc;
+  text-align: center;
+  display: block;
+  padding: 20px 0;
 }
