--- src/pages/records/index.config.ts
@@ -1,4 +1,0 @@
-export default definePageConfig({
-  navigationBarTitleText: "记录",
-  enablePullDownRefresh: true,
-});
